% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meristems.R
\docType{data}
\name{meristems}
\alias{meristems}
\title{Sample RNA data of similarity between batches of 1000 cells of tomato meristem cells.}
\format{A simple square matrix.}
\usage{
data(meristems)
}
\description{
This is a simple matrix where each entry is the similarity (correlation) between a pair of
batches. Negative correlations were changed to zero to simplify the analysis.
}
\examples{
data(meristems)
\donttest{slanter::sheatmap(showrownames=F, show_colnames=F)}
}
\keyword{datasets}
