% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slanter.R
\name{reorder_frame}
\alias{reorder_frame}
\title{Reorder the rows of a frame.}
\usage{
reorder_frame(frame, order)
}
\arguments{
\item{frame}{A data frame to reorder the rows of.}

\item{order}{An array containing indices permutation to apply to the rows.}

\item{...}{No additional arguments are allowed.}
}
\value{
The data frame with the new row orders.
}
\description{
If you expect \code{data[order,]} to "just work", you haven't been using R for very long. It is
this sort of thing that makes me *hate* coding in R.
}
