% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slanter.R
\name{sheatmap}
\alias{sheatmap}
\title{Plot a heatmap with values as close to the diagonal as possible.}
\usage{
sheatmap(data, ..., annotation_col = NULL, annotation_row = NULL,
  order_rows = T, order_cols = T, same_order = F, cluster_rows = T,
  cluster_cols = T, clustering_distance_rows = "euclidian",
  clustering_distance_cols = "euclidian",
  clustering_method = "ward.D2", clustering_callback = NA)
}
\arguments{
\item{data}{A rectangular matrix}

\item{...}{Additional flags to pass to \code{pheatmap}.}

\item{annotation_col}{Optional data frame describing each column.}

\item{annotation_row}{Optional data frame describing each row.}

\item{order_rows}{Whether to reorder the rows.}

\item{order_cols}{Whether to reorder the columns.}

\item{same_order}{Whether to apply the same order to both rows and columns.}

\item{cluster_rows}{Whether to cluster the rows, or the clustering to use.}

\item{cluster_cols}{Whether to cluster the columns, or the clustering to use.}

\item{clustering_distance_rows}{The default method for computing row distances (by default, \code{euclidian}).}

\item{clustering_distance_cols}{The default method for computing column distances (by default, \code{euclidian}).}

\item{clustering_method}{The default method to use for clustering (by default, \code{ward.D2}).}

\item{clustering_callback}{Is not supported.}
}
\value{
Whatever \code{pheatmap} returns.
}
\description{
Given a matrix expressing the cross-similarity between two (possibly different) sets of entities,
this will reorder it to move the high values close to the diagonal, for better visualizations.
}
\details{
If you want to cluster your data, you have two options:

If you do not specify a clustering, and just request one (e.g., \code{cluster_rows=T}), then
\code{sheatmap} will invoke \code{oclust} to generate the "best" clustering that is also
compatible with the slanted order. This is the default.

Otherwise, \code{sheatmap} will preserve the clustering it was given (e.g.,
\code{cluster_rows=hclust(...)}. Note that any given #' clustering tree allows for multiple
orders, since in each node, the two sub-trees can be flipped, without modifying the tree. Thus,
\code{sheatmap} will use the "best" order that is compatible with the given clustering.

In addition, you can give this function any of the \code{pheatmap} flags, and it will just pass
them on. This allows full control over the diagram's features.

Note that \code{clustering_callback} is not supported. In addition, the default
\code{clustering_method} here is \code{'ward.D2'} instead of \code{'complete'}, since the only
methods supported by \code{oclust} are \code{'ward.D'} and \code{'ward.D2'}.
}
