% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meristems.R
\docType{data}
\name{meristems}
\alias{meristems}
\title{Sample RNA data of similarity between batches of 1000 cells of tomato meristem cells.}
\format{A simple square matrix.}
\usage{
data(meristems)
}
\description{
This is a simple matrix where each entry is the similarity (correlation) between a pair of
batches. Negative correlations were changed to zero to simplify the analysis.
}
\examples{
data(meristems)
similarity <- meristems
similarity[similarity < 0] = 0
slanter::sheatmap(meristems, order_data=similarity, show_rownames=FALSE, show_colnames=FALSE)
}
\keyword{datasets}
