% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slanter.R
\name{reorder_frame}
\alias{reorder_frame}
\title{Reorder the rows of a frame.}
\usage{
reorder_frame(frame, order)
}
\arguments{
\item{frame}{A data frame to reorder the rows of.}

\item{order}{An array containing indices permutation to apply to the rows.}
}
\value{
The data frame with the new row orders.
}
\description{
You'd expect \code{data[order,]} to "just work". It doesn't for data frames with a single column,
which happens for annotation data, hence the need for this function. Sigh.
}
\examples{
df <- data.frame(foo=c(1, 2, 3))
df[c(1,3,2),]
slanter::reorder_frame(df, c(1,3,2))
}
